sitebuilder/requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.29.*
sqlmodel==0.0.16
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
jinja2==3.1.*
python-multipart==0.0.9     # загрузка файлов
